@@ -5,9 +5,12 @@
   "dependencies": {
     "@craco/craco": "^7.1.0",
     "axios": "^1.7.2",
+    "chart.js": "^4.5.1",
     "crypto-browserify": "^3.12.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "firebase": "^12.6.0",
+    "react": "^18.2.0",
+    "react-chartjs-2": "^5.3.1",
+    "react-dom": "^18.2.0",
     "react-helmet": "^6.1.0",
     "react-router-dom": "^6.30.1",
     "xlsx": "^0.18.5"
